Move playbook dot menu items to controls
