+++ webapp/src/components/backstage/playbook_editor/controls.tsx
@@ -17,6 +17,9 @@ import {
     ContentCopyIcon,
     PencilOutlineIcon,
     AccountMultipleOutlineIcon,
+    StarOutlineIcon,
+    StarIcon,
+    LinkVariantIcon,
 } from '@mattermost/compass-icons/components';
 
 import {Tooltip, OverlayTrigger} from 'react-bootstrap';
@@ -56,6 +59,9 @@ import useConfirmPlaybookArchiveModal from '../archive_playbook_modal';
 import CopyLink from 'src/components/widgets/copy_link';
 import useConfirmPlaybookRestoreModal from '../restore_playbook_modal';
 import {usePlaybookMembership} from 'src/graphql/hooks';
+import {StyledDropdownMenuItem} from '../shared';
+import {copyToClipboard} from 'src/utils';
+import {useLHSRefresh} from '../lhs_navigation';
 
 type ControlProps = {
     playbook: {
@@ -314,6 +320,58 @@ export const JoinPlaybook = ({playbook: {id: playbookId}, refetch}: ControlProps
             />
             {formatMessage({defaultMessage: 'Join playbook'})}
         </PrimaryButtonLarger>
+    );
+};
+
+export const FavoritePlaybookMenuItem = (props: {isFavorite: boolean, toggleFavorite: () => void}) => {
+    const {formatMessage} = useIntl();
+
+    return (
+        <StyledDropdownMenuItem onClick={props.toggleFavorite}>
+            {props.isFavorite ? (
+                <><StarOutlineIcon size={18}/>{formatMessage({defaultMessage: 'Unfavorite'})}</>
+            ) : (
+                <><StarIcon size={18}/>{formatMessage({defaultMessage: 'Favorite'})}</>
+            )}
+        </StyledDropdownMenuItem>
+    );
+};
+
+export const CopyPlaybookLinkMenuItem = (props: {playbookId: string}) => {
+    const {formatMessage} = useIntl();
+    const {add: addToast} = useToaster();
+
+    return (
+        <StyledDropdownMenuItem
+            onClick={() => {
+                copyToClipboard(getSiteUrl() + '/playbooks/playbooks/' + props.playbookId);
+                addToast(formatMessage({defaultMessage: 'Copied!'}));
+            }}
+        >
+            <LinkVariantIcon size={18}/>
+            <FormattedMessage defaultMessage='Copy link'/>
+        </StyledDropdownMenuItem>
+    );
+};
+
+export const LeavePlaybookMenuItem = (props: {playbookId: string}) => {
+    const currentUserId = useSelector(getCurrentUserId);
+    const refreshLHS = useLHSRefresh();
+
+    const {leave} = usePlaybookMembership(props.playbookId, currentUserId);
+    return (
+        <StyledDropdownMenuItem
+            onClick={async () => {
+                await leave();
+                refreshLHS();
+            }}
+        >
+            <CloseIcon
+                size={18}
+                color='currentColor'
+            />
+            <FormattedMessage defaultMessage='Leave'/>
+        </StyledDropdownMenuItem>
     );
 };
