+++ webapp/src/components/sidebar/create_playbook_dropdown.tsx
@@ -111,7 +111,7 @@ const CreatePlaybookDropdown = (props: CreatePlaybookDropdownProps) => {
     };
 
     return (
-        <Dropdown stopPropagationOnToggle={true}>
+        <Dropdown>
             <OverlayTrigger
                 delay={OVERLAY_DELAY}
                 placement='top'
